+++ spec/rspec/mocks/partial_double_spec.rb
@@ -3,6 +3,12 @@ module RSpec
     describe "A partial double" do
       let(:object) { Object.new }
 
+      it 'does not create an any_instance recorder when a message is allowed' do
+        expect {
+          allow(object).to receive(:foo)
+        }.not_to change { RSpec::Mocks.space.any_instance_recorders }.from({})
+      end
+
       it "names the class in the failure message" do
         expect(object).to receive(:foo)
         expect do
@@ -21,6 +27,20 @@ module RSpec
         object.instance_exec { @options = Object.new }
         expect(object).to receive(:blah)
         object.blah
+      end
+
+      it 'allows `class` to be stubbed even when `any_instance` has already been used' do
+        # See https://github.com/rspec/rspec-mocks/issues/687
+        # The infinite recursion code path was only triggered when there were
+        # active any instance recorders in the current example, so we make one here.
+        allow_any_instance_of(Object).to receive(:bar).and_return(2)
+
+        expect(object.class).not_to eq(String)
+        allow(object).to receive_messages(:foo => 1, :class => String)
+
+        expect(object.foo).to eq(1)
+        expect(object.class).to eq(String)
+        expect(object.bar).to eq(2)
       end
 
       it "can disallow messages from being received" do